requirements.txt
anthropic_bedrock==0.5.8
boto3==1.34.7
gradio==4.12.0
langchain==0.0.352
python-dotenv==1.0.0
